pyproject.toml
[project]
name = "ladrillo-gestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly-express>=0.4.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
]
